edge/requirements.txt
paho-mqtt>=1.6.1
pyserial>=3.5
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
tensorflow>=2.13.0
mediapipe>=0.10.0
